requirements.txt
playwright
behave
allure-behave
pytest
requests
